+++ crates/astria-sequencer-types/src/sequencer_block_data.rs
@@ -328,9 +328,7 @@ impl SequencerBlockData {
         }
 
         // generate the action tree root proof of inclusion in `Header::data_hash`
-        let hashed_txs = b.data.iter().map(|tx| sha256_hash(tx)).collect::<Vec<_>>();
-        let tx_tree = MerkleTree::from_leaves(hashed_txs);
-        let calculated_data_hash = tx_tree.root();
+        let (calculated_data_hash, tx_tree) = calculate_data_hash_and_tx_tree(&b.data);
         if calculated_data_hash != data_hash.as_bytes() {
             return Err(Error::CometBftDataHashReconstructedHashMismatch);
         }
@@ -360,6 +358,13 @@ impl SequencerBlockData {
         };
         Ok(data)
     }
+}
+
+fn calculate_data_hash_and_tx_tree(txs: &[Vec<u8>]) -> ([u8; 32], MerkleTree) {
+    let hashed_txs = txs.iter().map(|tx| sha256_hash(tx)).collect::<Vec<_>>();
+    let tx_tree = MerkleTree::from_leaves(hashed_txs);
+    let calculated_data_hash = tx_tree.root();
+    (calculated_data_hash, tx_tree)
 }
 
 fn sha256_hash(data: &[u8]) -> Vec<u8> {
@@ -521,5 +526,22 @@ mod test {
         let last_commit_hash = crate::calculate_last_commit_hash(&commit);
         assert!(matches!(last_commit_hash, Hash::Sha256(_)));
         assert!(expected_last_commit_hash.as_bytes() == last_commit_hash.as_bytes());
+    }
+
+    #[test]
+    fn calculate_data_hash() {
+        use base64::{
+            engine::general_purpose::STANDARD,
+            Engine as _,
+        };
+
+        // data_hash is calculated from the txs in a block, where the leaves of the merkle tree are
+        // the sha256 hashes of the txs
+        let tx = STANDARD.decode("CscBCsQBCg0vU2VxdWVuY2VyTXNnErIBCghldGhlcmV1bRJ4Avh1ggU5gIRZaC8AhQUD1cTyglIIlBtwp0/22gQLMRmQwVX9/9u8AvfuiA3gtrOnZAAAgMABoLnRqksJblEaolE6wbsAHYTAiSlA14+B5nvWuFrIfevnoBg+UGcWLC4eg1lZylqLnrL8okBc3vTS4qOO/J5sRtVDGixtZXRybzFsbDJobHAzM3J4eTdwN2s2YXhoeDRjdnFtdGcwY3hkZjZnemY5ahJ0Ck4KRgofL2Nvc21vcy5jcnlwdG8uc2VjcDI1NmsxLlB1YktleRIjCiEDJ/LvaMZTBcGX66geJOEmTm/fyyPTZKMUJoDtMDUmSPkSBAoCCAESGAoQCgV1dGljaxIHMTAwMDAwMBCAlOvcAyIIZXRoZXJldW0aQMhoTCUr84xgTkYxsFWDfHH2k+oHCPsKvbTpz8m5YrHfYMv6gdou6V8oj1v0B9ySD5VjMXQi1kJ9DZN6wD2buo8=").unwrap();
+        let expected_hash = STANDARD
+            .decode("rRDu3aQf1V37yGSTdf2fv9GSPeZ6/p0wJ9pjBl8IqFc=")
+            .unwrap();
+        let (data_hash, _) = super::calculate_data_hash_and_tx_tree(&[tx]);
+        assert_eq!(data_hash.to_vec(), expected_hash);
     }
 }
